producers/neo4j/requirements.txt
kafka-python==2.0.2
neo4j==4.4.1